Use column headers in xlsx file for Originnames
Modified run_boot_gsi_analysis to require that the user input
the column at which the stock group designations start in the
xlsx file (or the data frame).  The function that just gets
the Originnames from the column names.

Had to modify the tests to include that start column too.